Deprecate history integration's statistics API
