+++ homeassistant/components/recorder/websocket_api.py
@@ -1,13 +1,17 @@
 """The Recorder websocket API."""
 from __future__ import annotations
 
+from datetime import datetime as dt
 import logging
+from typing import Literal
 
 import voluptuous as vol
 
 from homeassistant.components import websocket_api
+from homeassistant.components.websocket_api import messages
 from homeassistant.core import HomeAssistant, callback, valid_entity_id
 from homeassistant.helpers import config_validation as cv
+from homeassistant.helpers.json import JSON_DUMP
 from homeassistant.util import dt as dt_util
 
 from .const import MAX_QUEUE_BACKLOG
@@ -15,6 +19,7 @@ from .statistics import (
     async_add_external_statistics,
     async_import_statistics,
     list_statistic_ids,
+    statistics_during_period,
     validate_statistics,
 )
 from .util import async_migration_in_progress, async_migration_is_live, get_instance
@@ -25,15 +30,125 @@ _LOGGER: logging.Logger = logging.getLogger(__package__)
 @callback
 def async_setup(hass: HomeAssistant) -> None:
     """Set up the recorder websocket API."""
-    websocket_api.async_register_command(hass, ws_validate_statistics)
-    websocket_api.async_register_command(hass, ws_clear_statistics)
-    websocket_api.async_register_command(hass, ws_get_statistics_metadata)
-    websocket_api.async_register_command(hass, ws_update_statistics_metadata)
-    websocket_api.async_register_command(hass, ws_info)
-    websocket_api.async_register_command(hass, ws_backup_start)
-    websocket_api.async_register_command(hass, ws_backup_end)
     websocket_api.async_register_command(hass, ws_adjust_sum_statistics)
+    websocket_api.async_register_command(hass, ws_backup_end)
+    websocket_api.async_register_command(hass, ws_backup_start)
+    websocket_api.async_register_command(hass, ws_clear_statistics)
+    websocket_api.async_register_command(hass, ws_get_statistics_during_period)
+    websocket_api.async_register_command(hass, ws_get_statistics_metadata)
+    websocket_api.async_register_command(hass, ws_list_statistic_ids)
     websocket_api.async_register_command(hass, ws_import_statistics)
+    websocket_api.async_register_command(hass, ws_info)
+    websocket_api.async_register_command(hass, ws_update_statistics_metadata)
+    websocket_api.async_register_command(hass, ws_validate_statistics)
+
+
+def _ws_get_statistics_during_period(
+    hass: HomeAssistant,
+    msg_id: int,
+    start_time: dt,
+    end_time: dt | None = None,
+    statistic_ids: list[str] | None = None,
+    period: Literal["5minute", "day", "hour", "month"] = "hour",
+) -> str:
+    """Fetch statistics and convert them to json in the executor."""
+    return JSON_DUMP(
+        messages.result_message(
+            msg_id,
+            statistics_during_period(hass, start_time, end_time, statistic_ids, period),
+        )
+    )
+
+
+async def ws_handle_get_statistics_during_period(
+    hass: HomeAssistant, connection: websocket_api.ActiveConnection, msg: dict
+) -> None:
+    """Handle statistics websocket command."""
+    start_time_str = msg["start_time"]
+    end_time_str = msg.get("end_time")
+
+    if start_time := dt_util.parse_datetime(start_time_str):
+        start_time = dt_util.as_utc(start_time)
+    else:
+        connection.send_error(msg["id"], "invalid_start_time", "Invalid start_time")
+        return
+
+    if end_time_str:
+        if end_time := dt_util.parse_datetime(end_time_str):
+            end_time = dt_util.as_utc(end_time)
+        else:
+            connection.send_error(msg["id"], "invalid_end_time", "Invalid end_time")
+            return
+    else:
+        end_time = None
+
+    connection.send_message(
+        await get_instance(hass).async_add_executor_job(
+            _ws_get_statistics_during_period,
+            hass,
+            msg["id"],
+            start_time,
+            end_time,
+            msg.get("statistic_ids"),
+            msg.get("period"),
+        )
+    )
+
+
+@websocket_api.websocket_command(
+    {
+        vol.Required("type"): "recorder/statistics_during_period",
+        vol.Required("start_time"): str,
+        vol.Optional("end_time"): str,
+        vol.Optional("statistic_ids"): [str],
+        vol.Required("period"): vol.Any("5minute", "hour", "day", "month"),
+    }
+)
+@websocket_api.async_response
+async def ws_get_statistics_during_period(
+    hass: HomeAssistant, connection: websocket_api.ActiveConnection, msg: dict
+) -> None:
+    """Handle statistics websocket command."""
+    await ws_handle_get_statistics_during_period(hass, connection, msg)
+
+
+def _ws_get_list_statistic_ids(
+    hass: HomeAssistant,
+    msg_id: int,
+    statistic_type: Literal["mean"] | Literal["sum"] | None = None,
+) -> str:
+    """Fetch a list of available statistic_id and convert them to json in the executor."""
+    return JSON_DUMP(
+        messages.result_message(msg_id, list_statistic_ids(hass, None, statistic_type))
+    )
+
+
+async def ws_handle_list_statistic_ids(
+    hass: HomeAssistant, connection: websocket_api.ActiveConnection, msg: dict
+) -> None:
+    """Fetch a list of available statistic_id."""
+    connection.send_message(
+        await get_instance(hass).async_add_executor_job(
+            _ws_get_list_statistic_ids,
+            hass,
+            msg["id"],
+            msg.get("statistic_type"),
+        )
+    )
+
+
+@websocket_api.websocket_command(
+    {
+        vol.Required("type"): "recorder/list_statistic_ids",
+        vol.Optional("statistic_type"): vol.Any("sum", "mean"),
+    }
+)
+@websocket_api.async_response
+async def ws_list_statistic_ids(
+    hass: HomeAssistant, connection: websocket_api.ActiveConnection, msg: dict
+) -> None:
+    """Fetch a list of available statistic_id."""
+    await ws_handle_list_statistic_ids(hass, connection, msg)
 
 
 @websocket_api.websocket_command(
